feat: experiment

+++ scripts/encode_pcd_with_bps.py
@@ -8,7 +8,7 @@ import rospy
 import open3d as o3d
 import torch
 import os
-
+from time import time
 
 class BPSEncoder():
     def __init__(self, client, bps_path=os.path.join(rospy.get_param('ffhnet_path'), 'models/basis_point_set.npy')):
@@ -30,20 +30,21 @@ class BPSEncoder():
 
         self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
 
-        self.VISUALIZE = True
+        self.VISUALIZE = False
 
     def handle_encode_pcd_with_bps(self, req, res):
+        start = time()
+        # Encode multi obj. This can not exit, for ffhnet eval.
+        if os.path.exists(self.multi_pcd_path):
+            obj_pcd = o3d.io.read_point_cloud(self.multi_pcd_path)
+            obj_tensor = torch.from_numpy(np.asarray(obj_pcd.points))
+            obj_tensor.to(self.device)
 
-        # Encode multi obj
-        obj_pcd = o3d.io.read_point_cloud(self.multi_pcd_path)
-        obj_tensor = torch.from_numpy(np.asarray(obj_pcd.points))
-        obj_tensor.to(self.device)
+            enc_dict = self.multi_bps.encode(obj_tensor)
+            print('encode pcd with multi bps encoder')
 
-        enc_dict = self.multi_bps.encode(obj_tensor)
-        print('encode pcd with multi bps encoder')
-
-        enc_np = enc_dict['dists'].cpu().detach().numpy()
-        np.save(self.multi_enc_path, enc_np)
+            enc_np = enc_dict['dists'].cpu().detach().numpy()
+            np.save(self.multi_enc_path, enc_np)
 
         # Encode single obj
         obj_pcd = o3d.io.read_point_cloud(self.pcd_path)
@@ -65,6 +66,7 @@ class BPSEncoder():
             bps_pcd.colors = o3d.utility.Vector3dVector(0.3 * np.ones(self.bps_np.shape))
             o3d.visualization.draw_geometries([obj_pcd, bps_pcd])
 
+        print("time for bps encoding:", time() - start)
         return True
